Added a function to be used to address Issue 35 (and maybe for Issue 36) as well.
The function has not yet been integrated into the speech string generation (it should be the last step before the string is sent to the speech engine). I will leave that up to you. The function needs a simple JSON dictionary/hash to be provided to it. Ideally, this should be short and will generally be application specific.