Pronouncing working (issue 35). I use the following form for a JSON dictionary:
{
    "JSON": "testing",
    "key2": "replacement2"
}

Where the file is loaded at the loading of the page via a 'data-ez-pronounce' for lack of a better alternative ATM. Can be easily changed later.

So an HTML file showing for it works could be like this (for example):

<body data-ez data-ez-pronounce="hello.json">
<p>JSON data</p>
<p>json data</p>
<p>Alexander V The Great</p>
</body>

Something I noticed: 'JSON' is changed with 'Testing', but 'json' is not. Is this the way you want it Bern, or do you want to change your function to accommadate this (I'd assume so)?